requirements.txt
Flask
SQLAlchemy
PyMySQL
python-dotenv
gunicorn
cryptography
